refactor: refactor api account active

+++ app/databases/account_active.py
@@ -29,34 +29,32 @@ class AccountActiveDatabase(Database):
     @staticmethod
     async def get(category, **kwargs):
         token = kwargs.get("token")
-        if category == "by_token_web":
-            if data_account_active := AccountActiveModel.objects(
-                token_web=token
-            ).first():
-                return data_account_active
-        if category == "by_token_email":
-            if data_account_active := AccountActiveModel.objects(
-                token_email=token
-            ).first():
-                return data_account_active
-
-    @staticmethod
-    async def delete(category, **kwargs):
-        token = kwargs.get("token")
-        user_id = kwargs.get("user_id")
         created_at = kwargs.get("created_at")
         if category == "by_token_web":
             if data_account_active := AccountActiveModel.objects(
                 token_web=token
             ).first():
-                data_account_active.delete()
-                return data_account_active
+                if data_account_active.expired_at > int(created_at):
+                    return data_account_active
+                else:
+                    data_account_active.user.updated_at = int(created_at)
+                    data_account_active.user.save()
+                    data_account_active.delete()
         if category == "by_token_email":
             if data_account_active := AccountActiveModel.objects(
                 token_email=token
             ).first():
-                data_account_active.delete()
-                return data_account_active
+                if data_account_active.expired_at > int(created_at):
+                    return data_account_active
+                else:
+                    data_account_active.user.updated_at = int(created_at)
+                    data_account_active.user.save()
+                    data_account_active.delete()
+
+    @staticmethod
+    async def delete(category, **kwargs):
+        user_id = kwargs.get("user_id")
+        created_at = kwargs.get("created_at")
         if category == "user_active_by_token_email":
             if user_data := UserModel.objects(id=user_id).first():
                 if data_account_active := AccountActiveModel.objects(
